refactor(agent-worker): extract runAISession into session-runner
Split the 2100-line worker.ts by moving the ~1000-line runAISession body and
its pure helpers into openclaw/session-runner.ts; OpenClawWorker.runAISession
is now a thin wrapper. Re-exports preserve every import path. Behavior
unchanged (worker suite: 550 pass).
